# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Takas Backend API",
    "description": "İkinci el takas mobil uygulaması için REST API",
    "version": "1.0.0",
    "contact": {
      "name": "Takas Backend Team",
      "email": "info@takas.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "title": {"type": "string"},
          "description": {"type": "string"},
          "price": {"type": "number"},
          "category_id": {"type": "integer"},
          "owner_id": {"type": "integer"},
          "condition": {"type": "string"},
          "status": {"type": "string"},
          "images": {"type": "array", "items": {"type": "string"}},
          "location": {"type": "string"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Ürünleri listele",
        "description": "Filtreleme ve sayfalama ile ürünleri listeler",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer"}, "description": "Sayfa numarası"},
          {"name": "per_page", "in": "query", "schema": {"type": "integer"}, "description": "Sayfa başına ürün sayısı"},
          {"name": "category_id", "in": "query", "schema": {"type": "integer"}, "description": "Kategori ID'si"},
          {"name": "search", "in": "query", "schema": {"type": "string"}, "description": "Arama terimi"}
        ],
        "responses": {
          "200": {
            "description": "Ürün listesi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {"type": "array", "items": {"$ref": "#/components/schemas/Product"}},
                    "total": {"type": "integer"},
                    "pages": {"type": "integer"},
                    "current_page": {"type": "integer"}
                  }
                }
              }
            }
          },
          "500": {"description": "Sunucu hatası"}
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Yeni ürün oluştur",
        "description": "Yeni ürün ekler",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "category_id"],
                "properties": {
                  "title": {"type": "string", "example": "iPhone 12"},
                  "description": {"type": "string", "example": "Mint condition iPhone 12"},
                  "price": {"type": "number", "example": 5000},
                  "category_id": {"type": "integer", "example": 1},
                  "condition": {"type": "string", "example": "Used"},
                  "location": {"type": "string", "example": "Istanbul"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ürün başarıyla oluşturuldu",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "product": {"$ref": "#/components/schemas/Product"}
                  }
                }
              }
            }
          },
          "400": {"description": "Geçersiz veri"},
          "401": {"description": "Yetkilendirme gerekli"},
          "500": {"description": "Sunucu hatası"}
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Ürün detayları",
        "description": "Belirli bir ürünün detaylarını döner",
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Ürün ID'si"}
        ],
        "responses": {
          "200": {
            "description": "Ürün detayları",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {"$ref": "#/components/schemas/Product"}
                  }
                }
              }
            }
          },
          "404": {"description": "Ürün bulunamadı"}
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Ürün güncelle",
        "description": "Mevcut ürünü günceller",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Ürün ID'si"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "description": {"type": "string"},
                  "price": {"type": "number"},
                  "category_id": {"type": "integer"},
                  "condition": {"type": "string"},
                  "location": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ürün başarıyla güncellendi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "product": {"$ref": "#/components/schemas/Product"}
                  }
                }
              }
            }
          },
          "403": {"description": "Yetkisiz erişim"},
          "404": {"description": "Ürün bulunamadı"}
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Ürün sil",
        "description": "Ürünü siler",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "product_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Ürün ID'si"}
        ],
        "responses": {
          "200": {"description": "Ürün başarıyla silindi"},
          "403": {"description": "Yetkisiz erişim"},
          "404": {"description": "Ürün bulunamadı"}
        }
      }
    },
...
  },
...
}